ex013

+++ exercicios/ex007.py
@@ -1,4 +1,4 @@
-n1 = int(input('Digite sua primeira nota:'))
-n2 = int(input('Digite sua segunda nota:'))
+n1 = float(input('Digite sua primeira nota:'))
+n2 = float(input('Digite sua segunda nota:'))
 media = (n1+n2)/2
 print('A média de suas notas é {}'.format(media))
